% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert_atlantis_output}
\alias{convert_atlantis_output}
\title{Load Atlantis outputfiles (netcdf)}
\usage{
convert_atlantis_output(model_path, filename, select_groups, select_variable,
  biomasspools, aggregate_layers = T)
}
\arguments{
\item{model_path}{Path of the ATLANTIS directory}

\item{filename}{Name of the ouputfile which shall be read in}

\item{select_variable}{ATLANTIS variables which shall be loaded! Only one variable can be loaded at a time! "Only 'N', 'Nums', 'ResN' and 'StructN'}

\item{aggregate_layers}{Specifies of layers shall be aggregated or not? Currently only T is available.}

\item{select_groups}{Vector of groups for which ouput shall be loaded. Should be derived from "functionalGroups.csv" using the column Name.}

\item{biomasspools}{Vector of biomasspool names as given in functionalGroups.csv!}
}
\value{
Dataframe in long format with the following informations: Species, timestep, polygon, agecl and value
}
\description{
This function loads Atlantis outputfiles (netcdf)
}
\details{
This functions converts the ATLANTIS output to a dataframe which can be processed in R.
}
\examples{
load_atlantis_output(model_path = "Z://Atlantis//ATLANTIS NSmodel", filename = "outputNorthSea.nc", select_variable = "ResN", biomasspools = c("large_crabs", "small_epifauna", "sessile_epifauna", "epifaunal_macrobenthos"))
}
\keyword{gen}

