% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_linf_k_fishbase}
\alias{extract_linf_k_fishbase}
\title{Extract values for Linf and k from www.fishbase.org}
\usage{
extract_linf_k_fishbase(fish, exclude_subspecies = T)
}
\arguments{
\item{vector}{of fish species with genus and species}

\item{specification}{if subspecies (e.g. Sprattus sprattus balticus) should be excluded!}
}
\value{
a dataframe with species, country, locality, linf and k!
}
\description{
This function extracts values for Linf and k from www.fishbase.org
}
\details{
Before the actual extraction takes place fishbaseh IDs for every species are extracted using the function "get_ids_fishbase". The IDs are needed to generate the URLs lateron. At the moment subspecies can only be excluded from the extraction.
}
\examples{
extract_linf_k_fishbase(c("Gadus morhua", "Merlangius merlangus"))
}

